Client: Moved ShoppingListDetails jsx into own component, renamed some things
