+++ client/src/features/shopping-lists/assets/shopping-list-detail-view.css
