+++ client/src/features/shopping-lists/index.js
@@ -1,5 +1,9 @@
-export {
-  createShoppingList,
-  getAllShoppingLists,
-  getSingleShoppingList,
-} from "./services/shoppingListFetches";
+export { ShoppingListCreator } from "./components/ShoppingListCreator.jsx";
+export { ShoppingListDetailView } from "./components/ShoppingListDetailView.jsx";
+export { ShoppingListsCollection } from "./components/ShoppingListsCollection.jsx";
+
+export { createShoppingList } from "./services/shoppingListFetches.js";
+export { getAllShoppingLists } from "./services/shoppingListFetches.js";
+export { getSingleShoppingList } from "./services/shoppingListFetches.js";
+export { editShoppingList } from "./services/shoppingListFetches.js";
+export { deleteShoppingList } from "./services/shoppingListFetches.js";
